Fix fact gathering when using --limit
If an operator wants to deploy a single node (e.g. add an additional
compute), they may want to use the --limit feature of Ansible to avoid
waiting for every play across every node.

The problem with this is that --limit will gather facts for that node
only, causing template errors such as the infamous ('dict object' has no
attribute u'ansible_eth0'").

Ansible has catered to this problem to an extent with it's
"delegate_facts" mechanism. The only problem is that in the default 'all
nodes' case, we end up with a storm of SSHs as each node SSHs to every
other. I'm solving this with a separate task to only use this mechanism
when a subset of nodes has been specified (see ansible_play_batch).

Useful links on this subject:
- https://medium.com/@george.shuklin/perfecting-forced-fact-gathering-in-ansible-1611f9c8d0d5#.tr5zs3e7x
- http://docs.ansible.com/ansible/playbooks_delegation.html#delegated-facts

Change-Id: Ibb691eae15cacd9e5129ae9280fd296f5ba95940
Closes-Bug: 1642004