Update Barbican to use correct mysql connection string
PyMySQL is prefered to PythonMySQL for Sqlalchemy, as it provides
python3 support and is actively maintained, and is therefore the
currently recommended lib for db connections.
 * https://wiki.openstack.org/wiki/PyMySQL_evaluation

Kolla currently uses PyMySQL for all connections bar Barbican
(which works fine with PyMySQL): once this commit is merged it will
be possible to remove the PythonMySQL libs, and mysql libs for kolla
images (except kolla-toolbox).

TrivialFix

Change-Id: Id256387134ca551a181c5e49c9b6d63f62b72523