add iscsi_ip_address to cinder.conf
rovider_location is wrong if we have multiple interfaces on storage-node, even if the storage-interface is allready configured.

For example we have an mgmt-interface with default-gw and an cluster-interface (bond0). We set:

network_interface: "bond0"

in global.yml which configure storage_interface correctly.

ansible/group_vars/all.yml:storage_interface: "{{ network_interface }}"

If the property iscsi_ip_address is not set, the provider_location points to the mgmt-ip, but this is the wrong network and the tgtd is not listening on that ip.