Fix bootstrap.yml in Bifrost
The bootstrap.yml playbook attempts to source
/opt/stack/ansible/hacking/env-setup and fails because no such file
exists (ansible is not installed in the location /opt/stack/). This step
is not required since the command bash ./scripts/env-setup.sh does not
install ansible from source in the Bifrost container as ansible is
already installed in the base image used by the Bifrost container.

Change-Id: Ibc26b4b77866127e7b57a30d80cd15734dedd85d
Closes-Bug: #1658741