Deployment Guide
This builds the deployment guide that will be published to
https://docs.openstack.org/project-deploy-guide/newton/.

In this approach, all existing documents in kolla-ansible/doc directory
will remain. The kolla-ansible/deploy-guide directory will simply
include all documents that exist in the kolla-ansible/doc directory.
All new changes will be made to the documents in the docs directory
which will be imported to the deployment-guide. However, I believe we
will also have to build the deployment-guide every time a change is
merged to the docs in the doc directory.

Change-Id: I98db4ec45403cf48086381b45bf5a12bbf998911