Support kolla-ansible installed in a virtualenv
Currently it is not possible to use kolla-ansible when it has been
installed in a virtualenv. Virtualenv-based installation may be
desirable when a suitable system package (e.g. RPM, .deb) is not
available, as it allows the user to install kolla-ansible and its
dependencies without affecting the system-wide packages.

This change checks for the presence of the $VIRTUAL_ENV environment
variable to determine whether we are running in an virtualenv. Since
kolla-ansible is not a python script, this comes with the caveat that we
must source the virtualenv activation script ($VIRTUAL_ENV/bin/activate)
before executing kolla-ansible.

Change-Id: I19efb96605bee3b16df3a0b78cd681e5f15499f6
Closes-Bug: #1530319