ironic-dnsmasq should use ironic-inspector group
The ironic_dnsmasq container is currently tied to the ironic-conductor
Ansible group. It is required only for Ironic inspector and should
really either be tied to the ironic-inspector group or have a new
ironic-dnsmasq group defined for it. This problem means that if all
hosts are removed from the ironic-inspector group I will still have an
ironic_dnsmasq container deployed.

This change uses the ironic-inspector group to determine where to place
the ironic_dnsmasq container.

Change-Id: I6af3f402795107b8b9d7a1619722f12cbf496257
Closes-Bug: #1666982