Remove heka_socket vol and unwedge some containers
The presence of heka_socket:/var/lib/kolla/heka with
containers that log to /dev/log, such as nova-ssh cause a
wedge on starting as /var/lib/kolla/heka/log is never
created due to the removal of heka.

This means that ssh data, such as config and keys are never
sync'd and sshd is never started.

Change-Id: Ia561526e6caf82eebd18c6e31cbeb1738b9ff602
Closes-Bug: #1672207
Signed-off-by: Dave Walker (Daviey) <email@daviey.com>