[deploy-guide] Moving the quick start to the deploy folder
Currently the project-deploy-guide links to the entire dev
doc repo for Kolla-Ansible. This is incorrect. It should link
only to installation information that can be found in the quick
start guide. This patch remedies that change, and ensures that
there is a link to the kolla deploy guide published at
docs.o.o.

Change-Id: I43738e341e996cb5a589a09c4c59c0dc8b058458