Revert "Remove useless variable glance_backend_file"
Turns out this variable is not so useless...! In
I2c167bae6f34a102cd7fe66401d7defe13bf2bf7 we need to determine if a file
based backend is in use in order to determine how many api servers to
start. Without this variable these kind of checks would all need to be
updated every time a new shared storage backend is added.

This reverts commit 867c3bda391626e3d4d6d4d6e803f08621b8f579.

Change-Id: I335acc12b68b577b33d6cb28675225c40e6231cd