Temporaly fix deploy gate
Nova service-list is returning empty registered service
when in database they are. Because of this, simple_cell_setup
is not executed and deploy stucks waiting for nova-compute.

This change temporaly checks the database for existing nova services
instead of openstack compute service list.

This change will need to be reverted once the command is fixed.

Change-Id: Ic508eb3ff03b5f233186353fc7697305cc792d14