Mixing binary and source images for A* and B* projects
In case Kolla's users want to deploy with both of
binary and source image, we should have a variable
install type that define install type for each project.

We also add specific image tag for each Openstack project.

This commit is implemented for Aodh, Barbican and Bifrost projects.

Change-Id: I926c9631c860b527941c40addc68896dd27bd1cc
Implements: blueprint mixing-binary-and-source-image