Mixing binary and source images for C* projects
In case Kolla's users want to deploy with both of
binary and source image, we should have a variable
install type that define install type for each project.

We also add specific image tag for each Openstack project.

This commit is implemented for Ceilometer, Ceph
Cinder, Cloudkitty and Congress projects.

Change-Id: I560ab84545db2e034abe03613c4e7cb407e64ab2
Implements: blueprint mixing-binary-and-source-image