Mixing binary and source images for M* projects
In case Kolla's users want to deploy with both of
binary and source image, we should have a variable
install type that define install type for each project.

We also add specific image tag for each Openstack project.

This commit is implemented for Magnum, Manila,
Mistral and Murano projects.

Change-Id: I6aaeb74f791c79229430a52b066327de7219a9b1
Implements: blueprint mixing-binary-and-source-image