Mixing binary and source images for T*, W*, Z* projects
In case Kolla's users want to deploy with both of
binary and source image, we should have a variable
install type that define install type for each project.

We also add specific image tag for each Openstack project.

This commit is implemented for Trove, Watcher, Zun projects.

Change-Id: I3fbf1e13064616a72cae195653f1c5fe4ea1d1b7
Implements: blueprint mixing-binary-and-source-image