Add evaluation_interval option
In ceilometer.conf there is a setting evaluation_interval.

The default is set to 60 seconds.

In pipeline.yaml,there is an interval set for those sources as well,
default value 300 seconds.

The evaluation_interval must be set >= the source interval in pipeline.yaml
or else when the evaluator runs it wont find any recent data
and set the state to insufficient data.

see:
https://docs.openstack.org/ocata/config-reference/telemetry/alarming-config-options.html

Change-Id: I82f061d1affc5c3ade75496684fe66e17928e1f3
Closes-Bug: #1704328