Fix the task about --limit does not perform
Kolla-ansible supports the deployment with specified tags,
"--tags, -t <tags> Only run plays and tasks tagged with
these values".
However, when specifying a tag, if the tag of the task
is not "always" or different from the specified tag,
it will not be run.
This task "Gather facts for all hosts (if using --limit)"
is not executed when the --limit parameter is added and
deployment with specified tag.

Closes-bug: #1711266

Change-Id: If50db2718b765f1d65b5d79eb042b0d95775bafc