Add a note to help user avoid docker start failure
The current example for including mtu is not suitable for all docker
version and distros. Fox example, the latest docker-ce 17.06 does not
support docker "docker daemon" command. And also docker doesn't
support socket parameter "fd://" on rpm based systems. We need to add
a note to remind user not copy that configuration straightly.

Change-Id: I8f2e0697eefe89abbe7da29612a625c4094c954e