Enable use of --limit when using bootstrap-servers
The ansible/kolla-host.yml playbook failed when called with a --limit
flag.

This was due to the fact that only the facts of the subset of hosts
were gathered but the "Generate /etc/hosts for all of the nodes" task
is using the ansible_<IFACE>.ipv4.address variable of all hosts.

The code for this fix is copied from the site.yml playbook
where it was used for the same reason.

Change-Id: I7e70caf750506ad8ab36ec024bd9647dd733fbd0
Closes-Bug: #1713725