Perform bootstrap on Horizon upgrade
The value of horizon_backend_database defaults to the value of
'enable_murano'. If a user happens to have enabled/deployed Murano prior
to upgrading Horizon, Horizon will then get configured to use the
backend database, though the bootstrap will be missing and results in a
broken Horizon on upgrade.

Closes-Bug: #1714949
Change-Id: I37c99c93f2966915c38d4e1aeaa48152254d62ea