filesystem_store_datadir should not be set when external Ceph is used
When an external Ceph cluster is used, the "filesystem_store_datadir"
option is set in the "glance-api.conf" configuration file.

It should not, it's because of this condition:

  {% elif enable_ceph | bool and glance_backend_ceph | bool %}

When external Ceph cluster is used "enable_ceph" is false and
"glance_backend_ceph" is true.

glance_backend_ceph variable should be enought for this condition.

Change-Id: I2a2ab420727888cfd9fcbc4bd30a08410163b36e
Closes-Bug: #1718728