Add notes on docker version in quickstart and multinode guide
In the quickstart guide, the min version of docker is 1.10.0,
and the max version is none. But the `docker daemon` doesn't work
since v1.13.0. Please refer to:
https://docs.docker.com/engine/deprecated/#docker-daemon-subcommand
"docker daemon subcommand
Deprecated In Release: v1.13.0
Target For Removal In Release: v17.12
The daemon is moved to a separate binary (dockerd), and should
be used instead."
I think it's better to let the readers know it.

Change-Id: I92882a3860704acdc31ff0c89e21ab2793e0b268