Disable glance file backend with external Ceph
When using external Ceph cluster, glance_api container is only running
on one controller instead of running on all of them.

But when an external Ceph cluster is used, enable_ceph is false so
glance_backend_file is true

Using "not glance_backend_ceph | bool" instead of "not enable_ceph |
bool" should be enough.

Change-Id: I191e5ea7a0d51b0e19ed58a819c3ade5070f623d
Closes-Bug: #1722905