XenAPI: support XenAPI in nova configuration
When using XenAPI as the virt driver, we need some XenAPI
specific configure options. This commit is to ensure nova's
configuration can be generated correctly for XenAPI. It
includes the folowing changes:
* Add XenAPI as one of the valid options for nova_compute_virt_type;
* Configure some options when XenAPI is chosen as the virt type.

This commit does NOT include the host and [vnc] configurations which
will be covered by a following commit.

References:
* XenServer (and other XAPI based Xen variants):
  https://docs.openstack.org/nova/pike/admin/configuration/hypervisor-xen-api.html
* XenCenter HIMN plugin (adding HIMN network which is used by XenAPI driver to
  communicate with XenServer):
  https://github.com/citrix-openstack/xencenter-himn-plugin
* nova configuration options:
  https://docs.openstack.org/nova/pike/configuration/config.html

Change-Id: Id34d247ab78976627f8e70685f27470b254cd418
blueprint: xenserver-support