Catch an error when pool creation failed for ceph
Ceph mon have a configure about "mon_max_pg_per_osd"
and default value is 300.
When pgs per osd is exceeded, pool creation will be failed.
But kolla-ansible ignoring this, so user may difficult to find
a main reason for the failure.
We need to catch an error when pool creation is failed.

Change-Id: I92493584faad9bc550fdeeb96b827537daafd0b8
Closes-Bug: #1747414