Add common_options to keystone container check
keystone_* containers are created via the kolla_docker ansible module
with common_options set to docker_common_options. However, when the
containers are checked, common_options are not passed to the
kolla_docker ansible module. This can cause the keystone_* containers
to be restarted during a reconfigure when there are no changes to
keystone configuration.

Add the common_options argument to the kolla_docker ansible module when
checking the keystone containers and set it to docker_common_options.

Change-Id: I44aefcf3d71faecaf1ffe384fd5a2f611e584a37
Closes-Bug: #1759922
(cherry picked from commit 735012c4bb4aaf83aa2eaadefde7b8b0a4674397)