Add task to create volumes path in cephfs
If using the cephfs native backend, when container start the
manila-share process with the manila user, no permissions are
prompted when creating the volume.
One way to solve the problem is to create the volumes folder in
advance in cephfs and modify the owner and group to 42429:42429,
which is manila's uid and gid.

Change-Id: I1ad63e1c4698ec8ee83461aafefa63041cfeb387
Closes-Bug: #1761108
(cherry picked from commit 6bfc6d4f3ac15269830b820e3a3223c0b506ca2a)