Reset multipath.conf
Current multipath.conf is just for IBM Storwize SVC model, not
generally applicable setting. With this nova-compute may encounter
error when execute 'multipath -l /dev/sda' because sda is defined
as blacklist in conf.

This patch reset the conf file back to default which is same to the
result of 'mpathconf --enable'. Vendor dependent parameters should
be configured separately.

Change-Id: I429e5a6e2a37d9db56d59f88cdbfab12db4813a8
Signed-off-by: Minho Ban <mhban@samsung.com>